rtpvrawpay: fix warning on macosx

+++ gst/rtp/gstrtpvrawpay.c
@@ -502,7 +502,8 @@ gst_rtp_vraw_pay_handle_buffer (GstBaseRTPPayload * payload, GstBuffer * buffer)
       if (!cont)
         break;
     }
-    GST_LOG_OBJECT (rtpvrawpay, "consumed %d bytes", outdata - headers);
+    GST_LOG_OBJECT (rtpvrawpay, "consumed %d bytes",
+        (gint) (outdata - headers));
 
     /* second pass, read headers and write the data */
     while (TRUE) {
